# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Transporte API",
        "description": "Documentación generada con Swagger para la API de transporte",
        "contact": {
            "email": "soporte@transporte.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor Local"
        },
        {
            "url": "https://tu-dominio.com",
            "description": "Servidor Producción"
        }
    ],
    "paths": {
...
        "/api/camiones": {
            "get": {
                "tags": [
                    "Camiones"
                ],
                "summary": "Listar camiones",
                "operationId": "73f14811afe9c701c13481e70ed0eeea",
                "responses": {
                    "200": {
                        "description": "Lista de camiones",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Camion"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Camiones"
                ],
                "summary": "Crear un nuevo camión",
                "operationId": "187f9820c013bc78fb0664f1382fcaf1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "placa",
                                    "marca",
                                    "modelo",
                                    "capacidad"
                                ],
                                "properties": {
                                    "placa": {
                                        "type": "string",
                                        "example": "ABC123"
                                    },
                                    "marca": {
                                        "type": "string",
                                        "example": "Volvo"
                                    },
                                    "modelo": {
                                        "type": "string",
                                        "example": "FH16"
                                    },
                                    "capacidad": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 15000
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Camión creado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Camion"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/camiones/{id}": {
            "get": {
                "tags": [
                    "Camiones"
                ],
                "summary": "Mostrar un camión específico",
                "operationId": "abd1d5be8d0a1c83142b580f84ac5238",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del camión",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles del camión",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Camion"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Camiones"
                ],
                "summary": "Actualizar un camión",
                "operationId": "0a2740d33d83659a12afd2d3f003264a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del camión",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "placa": {
                                        "type": "string",
                                        "example": "DEF456"
                                    },
                                    "marca": {
                                        "type": "string",
                                        "example": "Mercedes"
                                    },
                                    "modelo": {
                                        "type": "string",
                                        "example": "Actros"
                                    },
                                    "capacidad": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 20000
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Camión actualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Camion"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Camiones"
                ],
                "summary": "Eliminar un camión",
                "operationId": "dd10a49816bafccace92731c83b2a8fa",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del camión",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Camión eliminado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Camión eliminado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Camion": {
                "title": "Camion",
                "description": "Información de un camión",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "placa": {
                        "description": "Placa única del camión",
                        "type": "string",
                        "example": "ABC123"
                    },
                    "modelo": {
                        "description": "Modelo del camión",
                        "type": "string",
                        "example": "Volvo FH16"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-01-01T12:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-01-01T12:00:00Z"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}